# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/admin/stats": {
      "get": {
        "operationId": "AdminController_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get platform-wide statistics",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "operationId": "AdminController_listUsers",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all users",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/admin/properties/{id}/moderate": {
      "patch": {
        "operationId": "AdminController_moderate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModeratePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Moderate a property (change status)",
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/v1/auth/otp/request": {
      "post": {
        "operationId": "AuthController_requestOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestOtpDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "summary": "Request a 6-digit OTP via WhatsApp",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyOtpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Verify OTP and issue JWT tokens",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/properties/{id}/availability": {
      "get": {
        "operationId": "BookingsController_listAvailability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List availability blocks for a property",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings": {
      "post": {
        "operationId": "BookingsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a short-stay booking",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/my": {
      "get": {
        "operationId": "BookingsController_myBookings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the authenticated user's bookings",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/managed": {
      "get": {
        "operationId": "BookingsController_managed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List bookings on managed properties",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/v1/bookings/{id}/cancel": {
      "patch": {
        "operationId": "BookingsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a booking",
        "tags": [
          "Bookings"
        ]
      }
    },
    "/api/v1/leases": {
      "post": {
        "operationId": "LeasesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLeaseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a draft lease",
        "tags": [
          "Leases"
        ]
      }
    },
    "/api/v1/leases/managed": {
      "get": {
        "operationId": "LeasesController_managed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List leases on properties the user owns or manages",
        "tags": [
          "Leases"
        ]
      }
    },
    "/api/v1/leases/{id}/activate": {
      "patch": {
        "operationId": "LeasesController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate a draft lease (generates rent schedule)",
        "tags": [
          "Leases"
        ]
      }
    },
    "/api/v1/leases/{id}/schedule": {
      "get": {
        "operationId": "LeasesController_schedule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the rent schedule for a lease",
        "tags": [
          "Leases"
        ]
      }
    },
    "/api/v1/maintenance/tickets": {
      "post": {
        "operationId": "MaintenanceController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Open a maintenance ticket",
        "tags": [
          "Maintenance"
        ]
      },
      "get": {
        "operationId": "MaintenanceController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List tickets visible to the actor",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/v1/maintenance/tickets/my": {
      "get": {
        "operationId": "MaintenanceController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List tickets reported by the user",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/v1/maintenance/tickets/managed": {
      "get": {
        "operationId": "MaintenanceController_managed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List tickets on managed properties",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/v1/maintenance/tickets/{id}": {
      "patch": {
        "operationId": "MaintenanceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a ticket status or estimated cost",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/v1/maintenance/tickets/{id}/assign": {
      "patch": {
        "operationId": "MaintenanceController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a ticket to a technician",
        "tags": [
          "Maintenance"
        ]
      }
    },
...
  },
  "info": {
    "title": "Paradis Immo API",
    "description": "Hybrid real estate platform API — Congo (CG). Generated snapshot.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ModeratePropertyDto": {
        "type": "object",
        "properties": {}
      },
      "RequestOtpDto": {
        "type": "object",
        "properties": {}
      },
      "VerifyOtpDto": {
        "type": "object",
        "properties": {}
      },
...
      "CreateBookingDto": {
        "type": "object",
        "properties": {}
      },
      "CreateLeaseDto": {
        "type": "object",
        "properties": {}
      },
      "CreateTicketDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {}
      },
      "AssignTicketDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}